restructure common methods of base and hometemp
