+++ hometemp.py
@@ -5,56 +5,18 @@ import schedule
 import threading
 import time
 from datetime import datetime
-from gpiozero import CPUTemperature
 
 from core.command import CommandService
 from core.distribute import send_visualization_email
 from core.database import DwDDataHandler, GoogleDataHandler, UlmDeHandler, SensorDataHandler, WetterComHandler
-from core.sensors.dht import DHT, DHTResult
-from core.virtualization import PostgresDockerManager
+from core.sensors.dht import get_sensor_data
+from core.sensors.util import get_temperature
+from core.virtualization import init_postgres_container
 from core.plotting import draw_plots
 
 # GLOBAL Variables
 log = None
 command_service = None
-
-
-# ------------------------------- Raspberry Pi Temps ----------------------------------------------
-
-def get_temperature():
-    cpu = CPUTemperature()
-    return float(cpu.temperature)
-
-
-def get_sensor_data(used_pin):
-    """
-    Returns temperature and humidity data measures by AM2302 Sensor and the measurement timestamp.
-    """
-    max_tries = 15
-    tries = 0
-    DHT_SENSOR = DHT(used_pin, False)
-    while True:
-        if tries >= max_tries:
-            log.error(f"Failed to retrieve data from AM2302 sensor: Maximum retries reached.")
-            break
-        else:
-            time.sleep(2)
-            result = DHT_SENSOR.read()
-            if result.error_code == DHTResult.ERR_NOT_FOUND:
-                tries += 1
-                log.warning(f"({tries}/{max_tries}) Sensor could not be found. Using correct pin?")
-                # maybe exit here but invastige when this error occurs (should never)
-                continue
-            elif not result.is_valid():
-                tries += 1
-                log.warning(f"({tries}/{max_tries}) Sensor data invalid")
-                continue
-            elif result.is_valid() and result.error_code == DHTResult.ERR_NO_ERROR:
-                # postgres expects timestamp ins ISO 8601 format
-                timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
-                return result.temperature, result.humidity, timestamp
-
-    return None, None, None
 
 
 def _get__visualization_data():
@@ -139,7 +101,7 @@ def collect_and_save_to_db():
     handler = SensorDataHandler(auth['db_port'], auth['db_host'], auth['db_user'], auth['db_pw'], 'sensor_data')
     handler.init_db_connection()
     cpu_temp = get_temperature()
-    room_temp, humidity, timestamp = get_sensor_data(int(hometemp_config()["sensor_pin"]))
+    room_temp, humidity, timestamp = get_sensor_data(int(hometemp_config()["sensor_pin"]), False)
     if room_temp is not None and humidity is not None and timestamp is not None:
         log.info(
             "[Measurement {0}] CPU={1:f}*C, Room={2:f}*C, Humidity={3:f}%".format(timestamp, cpu_temp, room_temp,
@@ -149,26 +111,9 @@ def collect_and_save_to_db():
     log.info("Done")
 
 
-def init_postgres_container():
-    log.info("Checking for existing database")
-    auth = database_config()
-    docker_manager = PostgresDockerManager(auth["db_name"], auth["db_user"], auth["db_pw"])
-    container_name = auth["container_name"]
-    if docker_manager.container_exists(container_name):
-        log.info("Reusing existing database container")
-        return docker_manager.start_container(container_name)
-    else:
-        log.info("No database container found. Creating database container")
-        if docker_manager.pull_postgres_image():
-            docker_manager.create_postgres_container(container_name)
-            return docker_manager.start_container(container_name)
-
-    return False
-
-
 def main():
     log.info(f"------------------- HomeTemp v{hometemp_config()['version']} -------------------")
-    if not init_postgres_container():
+    if not init_postgres_container(database_config()):
         log.error("Postgres container startup error! Shutting down ...")
         exit(1)
         # after restart, database needs some time to start
